backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0

